==> notebooks/1. data_cleaning.ipynb <==
## Dataset overview:

import pandas as pd
import numpy as np
import functions

df = pd.read_csv("../data/raw/jobs_in_data.csv")
df.head()

df.shape

df.columns

## Unique values

functions.unique_values(df)

## Missing values

functions.nan_values(df)

## Duplicates

df.duplicated().sum()

df.drop_duplicates(keep="first", inplace=True)

df.shape

df

df.reset_index(drop=True, inplace=True)
df

## Data Types

df.dtypes

## Transformations

I only need a column with the salary information and I would like it to be in Euros, which is more suitable for my target audience.

df['salary_in_euros'] = df['salary_in_usd']*0.92
df.head()

df['salary_in_euros'] = df['salary_in_euros'].astype(int)
df.head()

df.drop(columns=['salary_currency','salary', 'salary_in_usd'], inplace=True)

df.head()

df.to_csv('../data/cleaned/jobs_in_data_cleaned_v1.csv', index=False)
