==> notebooks/2. data_cleaning_2.ipynb <==
## Data Overview

I would now like to add a new column to the original dataframe of jobs in data that accounts for the different living costs depending on the employee's residence because sometimes the absolute salary doesn't tell the whole story. With that in mind i researched a new dataframe with data regarding cost of living by country and I'm going to add it to the original dataframe.

import pandas as pd
import numpy as np
import functions

cost_of_living = pd.read_csv('../data/raw/cost-of-living-by-country-2024.csv')
cost_of_living.head()

functions.nan_values(cost_of_living)

This dataframe has different factors for cost of living from different sources. I decided to go for the Global Economy factor since it's a reliable source and doesn't have so many missing values that are relevant in my original dataframe.

cost_of_living_sliced = cost_of_living[['country','costOfLivingByCountry_cl_globalEcon']]

cost_of_living_sliced.rename (columns={'costOfLivingByCountry_cl_globalEcon':'cost_of_living'}, inplace=True)

cost_of_living_sliced

cost_of_living_sliced.to_csv('../data/cleaned/cost_of_living_cleanedV1.csv', index=False)
