==> notebooks/3. merging_dataframes.ipynb <==
## Merge dataframes

Now I want to add the column cost_of_living to my dataframe to add extra context to the salary values. The cost of living will be merged by country with the employee_residence.

import pandas as pd
import numpy as np
import functions

jobs_in_data = pd.read_csv('../data/cleaned/jobs_in_data_cleaned_v1.csv')
cost_of_living = pd.read_csv('../data/cleaned/cost_of_living_cleanedV1.csv')

jobs_in_data.head()

cost_of_living.head()

jobs_in_data_merged = jobs_in_data.merge(cost_of_living, left_on='employee_residence', right_on='country', how='left').drop(columns='country')

jobs_in_data_merged.head()

Now I want to add a column named "purchasing_power" that is the ratio between the employee's salary and the cost of living in his country of residence.

jobs_in_data_merged['purchasing_power'] = round(jobs_in_data_merged['salary_in_euros']/jobs_in_data_merged['cost_of_living'], 2)

jobs_in_data_merged

jobs_in_data_merged.isna().sum()

Since there were missing values in the cost of living dataframe, when I merged it with the jobs in data dataset, it created a few missing values in the final dataset . Thankfully, there are very few of them, so I can drop the missing rows without impacting the integrity of the data.

jobs_in_data_merged[jobs_in_data_merged.isna().any(axis=1)]

jobs_in_data_merged.dropna(inplace=True)

jobs_in_data_merged.reset_index(drop=True, inplace=True)

jobs_in_data_merged

jobs_in_data_merged.to_csv('../data/cleaned/jobs_in_data_merged.csv', index=False)
