==> notebooks/4. cardinality.ipynb <==
## Cardinality

Now I want to deal with cardinality in the column jobs_category

import pandas as pd
import numpy as np

df = pd.read_csv('../data/cleaned/jobs_in_data_merged.csv')
df.head()

df['job_category'].value_counts()

jobs_dict = {
    'Data Science and Research': 'Data Science',
    'Data Engineering': 'Data Engineering',
    'Machine Learning and AI': 'Data Science',
    'Data Analysis': 'Data Analysis',
    'Leadership and Management': 'Leadership and Management',
    'BI and Visualization': 'Data Analysis',
    'Data Architecture and Modeling': 'Data Engineering',
    'Data Management and Strategy': 'Data Engineering',
    'Data Quality and Operations': 'Data Engineering',
    'Cloud and Database': 'Data Engineering'
}

df['job_field'] = df['job_category'].map(jobs_dict)

df['job_field'].value_counts()

jobs_dict_2 = {
    "Data Science": ['Science', 'Scientist', 'AI', 'ML', 'Machine Learning', 'Deep Learning', 'Model'],
    'Data Engineering' : ['Engineer', 'Engineering', 'Architect', 'DB', 'Database', 'Azure', 'ETL'],
    'Data Analysis'  : ['Analyst', 'Analysis', 'Analytic', 'Business', 'BI'],
}

def assign_job_field(job_title):
    for key, values in jobs_dict_2.items():
        if any(word in job_title for word in values):
            return key
    return 'Other'

df['job_field'] = df.apply(
    lambda x: assign_job_field(x['job_title']) if x['job_category'] == 'Leadership and Management' else x['job_field'], 
    axis=1
)

df['job_field'].value_counts()

df['job_field'].unique()

df.to_csv('../data/cleaned/jobs_in_data_cardinality.csv', index=False)
