==> notebooks/5. EDA.ipynb <==
import pandas as pd
import numpy as np
import matplotlib as plt
import seaborn as sns

df = pd.read_csv('../data/cleaned/jobs_in_data_cardinality.csv')
df.head()

df['employee_residence'].value_counts()
