==> notebooks/6. cat_ordinal_columns.ipynb <==
## Categorical Ordinal columns

In this step I'm going to assume that there are categories in some of the features that have more weight for the final salary than others. In this case I assume those are categorical ordinal columns and I will replace the values with numbers to reflect that weight.

import pandas as pd
import numpy as np

df = pd.read_csv("../data/cleaned/jobs_in_data_cardinality.csv")

df.head()

#### Experience Level
I'm going to start by experience level assuming that the more experience you have the more well paid you are.

df['experience_level'].value_counts()

experience_level = {
    'Entry-level': 1,
    'Mid-level': 2,
    'Senior': 3,
    'Executive': 4,
}


df['experience_level'] = df['experience_level'].replace(experience_level)
df['experience_level'].value_counts()

#### Employment Type
I assume also that the type of employment will have a determinant weight on the salary because that considers the number of hours an employee works for the company.

df['employment_type'].value_counts()

employment_type = {
    'Full-time': 4,
    'Contract': 3,
    'Part-time': 2,
    'Freelance':1,
}


df['employment_type'] = df['employment_type'].replace(employment_type)
df['employment_type'].value_counts()

df['work_setting'].value_counts()

work_setting = {
    'In-person': 3,
    'Hybrid': 2,
    'Remote': 1,
}

df['work_setting'] = df['work_setting'].replace(work_setting)
df['work_setting'].value_counts()

df['company_size'].value_counts()

company_size = {
    'L': 3,
    'M': 2,
    'S': 1,
}

df['company_size'] = df['company_size'].replace(company_size)
df['company_size'].value_counts()

#### Other categories
After analyzing the other features I decided that there were no more categorical ordinal columns so I won't change anything else.

df.dtypes

df.to_csv("../data/cleaned/6.jobs_in_data.csv", index=False)
