==> notebooks/10. Ridge-final.ipynb <==
## Scaling and transformations
Now before proceeding to building the models I'm going to do the X/y split and scale/transform the data

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')

df = pd.read_csv('../data/cleaned/6.jobs_in_data.csv')
df.head()

## X/y Split
The target will be "salary_in_euros". I also want to drop the columns 'job_title' and 'job_category' because they are redundant for creating the model since I already added the column 'job_field' before with the categories that I want to work with.

X = df.drop(columns=['work_year', 'salary_in_euros', 'job_title', 'job_category', 'purchasing_power'], axis=1)
y = df['salary_in_euros']

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

display(X_train.shape)
display(X_test.shape)
display(y_train.shape)
display(y_test.shape)

## Dividing X into numerical and categorical

X_train_num = X_train.select_dtypes(np.number)
X_train_cat = X_train.select_dtypes(object)
X_test_num = X_test.select_dtypes(np.number)
X_test_cat = X_test.select_dtypes(object)

X_train_num.head()

X_test_num.head()

X_train_num.describe().T

## Power Transformer

#applying the power transformer (it will apply also automatically the Standard Scaler)
from sklearn.preprocessing import PowerTransformer
import pickle

transformer = PowerTransformer(method="box-cox")
transformer.fit(X_train_num)

path = "../ml/transformers/"
scaler_file_name = "power_transformer_x.pkl"

with open(path + scaler_file_name, "wb") as file:
    pickle.dump(transformer, file)

X_train_num_transformed = transformer.transform(X_train_num)
X_test_num_transformed = transformer.transform(X_test_num)

X_train_num_transformed_df = pd.DataFrame(X_train_num_transformed, columns=X_train_num.columns , index=X_train_num.index)
X_test_num_transformed_df = pd.DataFrame(X_test_num_transformed, columns=X_test_num.columns , index=X_test_num.index)

X_test_num_transformed_df.describe().T

X_train_num_transformed_df.head()

X_test_num_transformed_df.head()

## Encoding categorical features

X_train_cat.head()

#using the One Hot Encoder to encode the categorical nominal features
from sklearn.preprocessing import OneHotEncoder

encoder = OneHotEncoder(handle_unknown='ignore')
encoder.fit(X_train_cat)

path = "../ml/encoders/"
encoder_file_name = "one_hot_encoder.pkl"

with open(path + encoder_file_name, "wb") as file:
    pickle.dump(encoder, file)

X_train_cat_encoded = encoder.transform(X_train_cat).toarray()
X_test_cat_encoded = encoder.transform(X_test_cat).toarray()

encoded_feature_names = encoder.get_feature_names_out(X_train_cat.columns)

X_train_cat_encoded_df = pd.DataFrame(X_train_cat_encoded, columns=encoded_feature_names, index=X_train_cat.index)
X_test_cat_encoded_df = pd.DataFrame(X_test_cat_encoded, columns=encoded_feature_names, index=X_test_cat.index)

X_train_cat_encoded_df.head()

X_test_cat_encoded_df.head()

X_train_concat = pd.concat([X_train_num_transformed_df, X_train_cat_encoded_df], axis=1)
X_train_concat

X_test_concat = pd.concat([X_test_num_transformed_df, X_test_cat_encoded_df], axis=1)
X_test_concat

#applying the Min Max Scaler to the whole concatenated dataset
from sklearn.preprocessing import MinMaxScaler

scaler = MinMaxScaler()
scaler.fit(X_train_concat)

path = "../ml/scalers/"
scaler_file_name = "MinMaxScaler_concat.pkl"

with open(path + scaler_file_name, "wb") as file:
    pickle.dump(scaler, file)

X_train_concat_transformed = scaler.transform(X_train_concat)
X_test_concat_transformed = scaler.transform(X_test_concat)

X_train_concat_transformed_df = pd.DataFrame(X_train_concat_transformed, columns=X_train_concat.columns , index=X_train_concat.index)
X_test_concat_transformed_df = pd.DataFrame(X_test_concat_transformed, columns=X_test_concat.columns , index=X_test_concat.index)

X_train_concat_transformed_df.describe().T

#applying the power transformer on the target (y)
pt = PowerTransformer()

#fit transformer
y_train_2d = y_train.values.reshape(-1, 1)
y_test_2d = y_test.values.reshape(-1, 1)

pt.fit(y_train_2d)

#transform
y_train_trans = pt.transform(y_train_2d)
y_test_trans = pt.transform(y_test_2d)

#pickle
path = "../ml/transformers//"
transformer_file_name = "power_transformer_y.pkl"
with open(path + transformer_file_name, "wb") as file:
    pickle.dump(pt, file)


from sklearn.linear_model import Ridge
from sklearn.model_selection import GridSearchCV

ridge = Ridge()

alpha = [8, 12, 16]
max_iter = [15, 25, None]
tol = [0.001]
solver = ['auto']

grid = {'alpha': alpha,
        'max_iter': max_iter,
        'tol': tol,
        'solver': solver,
       }

grid_search = GridSearchCV(estimator = ridge, param_grid = grid, cv = 5)
grid_search.fit(X_train_concat_transformed_df, y_train_trans)


grid_search.best_params_

print("The best R2 for the best hyperparameters is {:.2f}".format(grid_search.best_score_))
print("The performance of the model using the best gridsearchcv hyperpameters is {:.2f}".format(grid_search.best_estimator_.score(X_test_concat_transformed_df, y_test_trans)))


#fitting the model
ridge = Ridge(alpha=8, max_iter=15, tol=000.1)

ridge.fit(X_train_concat_transformed_df, y_train_trans)

#pickle
path = "../ml/models//"
model_file_name = "ridge.pkl"
with open(path + model_file_name, "wb") as file:
    pickle.dump(ridge, file)

#predictions on train set
y_train_trans_pred = ridge.predict(X_train_concat_transformed_df)

#predictions on test set
y_test_trans_pred = ridge.predict(X_test_concat_transformed_df)

#score on the train set
print(round(ridge.score(X_train_concat_transformed_df, y_train_trans), 2))

#score on the test set
print(round(ridge.score(X_test_concat_transformed_df, y_test_trans), 2))

y_train_2d = y_train_trans_pred.reshape(-1, 1)
y_test_2d = y_test_trans_pred.reshape(-1, 1)

y_train_pred_array = pt.inverse_transform(y_train_2d).ravel()
y_test_pred_array = pt.inverse_transform(y_test_2d).ravel()

y_train_pred=pd.Series(y_train_pred_array)
y_test_pred=pd.Series(y_test_pred_array)
