==> notebooks/12. App.ipynb <==
# Lasso Model

## Scaling and transformations
Now before proceeding to building the model I'm going to do the X/y split and scale/transform the data

import pandas as pd
import numpy as np

df = pd.read_csv('../data/cleaned/6.jobs_in_data.csv')
df.head(5)


df['experience_level'].unique()

import streamlit as st
import pandas as pd
cost_of_living = pd.read_csv('../data/cleaned/cost_of_living_cleanedV1.csv')

def input_data_streamlit(cost_of_living_dataframe):
    # Define the columns and potential values for dropdowns
    columns = {
        'work_year': [2020, 2021, 2022, 2023, 2024, 2025],
        'employee_residence': ['Country1', 'Country2', 'Country3'],  # Example countries
        'experience_level': ['Junior', 'Mid', 'Senior', 'Lead'],
        'employment_type': ['Full-time', 'Part-time', 'Contract'],
        'work_setting': ['Office', 'Hybrid', 'Remote'],
        'company_location': ['Location1', 'Location2', 'Location3'],  # Example locations
        'company_size': ['Small', 'Medium', 'Large'],
        'salary_in_euros': None,  # For numerical input
        'cost_of_living': None,  # For numerical input
        'job_field': ['Data Science', 'Engineering', 'Product Management']
    }

    # Dictionary to store user inputs
    data = {}

    # Use Streamlit widgets to get user input
    for column, options in columns.items():
        if options:  # If options are provided, use a select box
            data[column] = st.selectbox(f"Select {column}:", options)
        else:  # For numerical input
            data[column] = st.number_input(f"Enter {column}:", step=1.0)

    # Button to add data to DataFrame
    if st.button('Add Entry'):
        new_row = pd.DataFrame([data], columns=columns.keys())
        return new_row

    df = pd.DataFrame(columns=columns.keys())  # return empty DataFrame if button not pressed
    df_merged = df.merge(cost_of_living_dataframe, left_on='employee_residence', right_on='country', how='left').drop(columns='country')

    return df_merged

# Use the function in Streamlit
st.title('Data Input for Regression Model')
user_df = input_data_streamlit(cost_of_living)

if not user_df.empty:
    st.write('Data Entered:')
    st.write(user_df)

jobs_in_data = pd.read_csv('../data/cleaned/jobs_in_data_cleaned_v1.csv')
cost_of_living = pd.read_csv('../data/cleaned/cost_of_living_cleanedV1.csv')
jobs_in_data_merged = jobs_in_data.merge(cost_of_living_dataframe, left_on='employee_residence', right_on='country', how='left').drop(columns='country')


df

## X/y Split
The target will be "salary_in_euros". I also want to drop the columns 'job_title' and 'job_category' because they are redundant for creating the model since I already added the column 'job_field' before with the categories that I want to work with.
(edit) After some trial and error testing I decided to drop the columns employee_residence and company_location since the high amount of unique values was influencing the performance of the model and causing overfitting. I dropped also purchasing power because that was a feature that I added from the target column for EDA purposes so it makes no sense in using it to predict the target value.

X = df.drop(columns=['job_title', 'salary_in_euros', 'purchasing_power'], axis=1)
y = df['salary_in_euros']

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

display(X_train.shape)
display(X_test.shape)
display(y_train.shape)
display(y_test.shape)

## Dividing X into numerical and categorical

X_train_num = X_train.select_dtypes(np.number)
X_train_cat = X_train.select_dtypes(object)
X_test_num = X_test.select_dtypes(np.number)
X_test_cat = X_test.select_dtypes(object)

X_train_num.head()

X_test_num.head()

X_train_num.describe().T

## Scaling numerical features
Since the numerical features have very different ranges I'm going to use the Standard Scaler.

from sklearn.preprocessing import MinMaxScaler
import pickle

scaler = MinMaxScaler()
scaler.fit(X_train_num)

path = "../ml/scalers/"
scaler_file_name = "MinMaxScaler.pkl"

with open(path + scaler_file_name, "wb") as file:
    pickle.dump(scaler, file)

X_train_num_transformed = scaler.transform(X_train_num)
X_test_num_transformed = scaler.transform(X_test_num)

# from sklearn.preprocessing import StandardScaler
# import pickle

# scaler = StandardScaler()
# scaler.fit(X_train_num)

# path = "../ml/scalers/"
# scaler_file_name = "standard_scaler.pkl"

# with open(path + scaler_file_name, "wb") as file:
#     pickle.dump(scaler, file)

# X_train_num_transformed = scaler.transform(X_train_num)
# X_test_num_transformed = scaler.transform(X_test_num)

X_train_num_transformed_df = pd.DataFrame(X_train_num_transformed, columns=X_train_num.columns , index=X_train_num.index)
X_test_num_transformed_df = pd.DataFrame(X_test_num_transformed, columns=X_test_num.columns , index=X_test_num.index)

X_train_num_transformed_df.head()

X_test_num_transformed_df.head()

## Encoding categorical features

X_train_cat.head()

from sklearn.preprocessing import OneHotEncoder

encoder = OneHotEncoder(handle_unknown='ignore')
encoder.fit(X_train_cat)

path = "../ml/encoders/"
encoder_file_name = "one_hot_encoder.pkl"

with open(path + encoder_file_name, "wb") as file:
    pickle.dump(encoder, file)

X_train_cat_encoded = encoder.transform(X_train_cat).toarray()
X_test_cat_encoded = encoder.transform(X_test_cat).toarray()

encoded_feature_names = encoder.get_feature_names_out(X_train_cat.columns)

X_train_cat_encoded_df = pd.DataFrame(X_train_cat_encoded, columns=encoded_feature_names, index=X_train_cat.index)
X_test_cat_encoded_df = pd.DataFrame(X_test_cat_encoded, columns=encoded_feature_names, index=X_test_cat.index)

X_train_cat_encoded_df.head()

X_test_cat_encoded_df.head()

X_train_concat = pd.concat([X_train_num_transformed_df, X_train_cat_encoded_df], axis=1)
X_train_concat

X_test_concat = pd.concat([X_test_num_transformed_df, X_test_cat_encoded_df], axis=1)
X_test_concat

from sklearn.model_selection import GridSearchCV

n_estimators = [80,100,120]
learning_rate = [0.01, 0.1, 1]
max_depth = [4, 6, 8]
min_samples_split = [3,5,7]
min_samples_leaf = [0.5,1,2]

grid = {'n_estimators': n_estimators,
        'learning_rate': learning_rate,
        'max_depth': max_depth,
        'min_samples_split': min_samples_split,
        'min_samples_leaf': min_samples_leaf,
       }


from sklearn.ensemble import GradientBoostingRegressor
import os

gb_reg = GradientBoostingRegressor()


grid_search = GridSearchCV(estimator = gb_reg, param_grid = grid, cv = 5)
grid_search.fit(X_train_concat, y_train)

grid_search.best_params_

print("The best R2 for the best hyperparameters is {:.2f}".format(grid_search.best_score_))
print("The performace of the Gradient Boost Regressor using the best gridsearchcv hyperpameters is {:.2f}".format(grid_search.best_estimator_.score(X_test_concat, y_test)))


import os

gb_reg = Gr(alpha=30, max_iter=100, selection='random', tol=0.1)

gb_reg.fit(X_train_concat, y_train)

path = "../ml/models/"
isExist = os.path.exists(path)
if not isExist:
    os.makedirs(path)
    print("The new directory is created!")

    filename = "Gradient Boost Regressor.pkl"
    with open(path+filename, "wb") as file:
        pickle.dump(lasso_reg, file)

y_train_pred = gb_reg.predict(X_train_concat)
y_test_pred  = gb_reg.predict(X_test_concat)

y_test_pred

import functions

functions.error_metrics_report(y_train, y_test, y_train_pred, y_test_pred)

### Conclusion:
